+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # NShell Changelog
 
+## Build 2
+
+- Now you can type something in, but then it just says goodbye.
+
 ## Build 1
 
 - Added build information